requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
gunicorn
mysqlclient
mysql-connector-python
